Separate out test_versioned_models.py
